fixed MSVC 2008 compile errors and warnings
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@2983 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ src/cv/cvsurf.cpp
@@ -670,7 +670,7 @@ cvExtractSURF( const CvArr* _img, const CvArr* _mask,
 			   int useProvidedKeyPts)
 {
     const int ORI_RADIUS = cv::SURFInvoker::ORI_RADIUS;
-    const int ORI_SIGMA = cv::SURFInvoker::ORI_SIGMA;
+    const float ORI_SIGMA = cv::SURFInvoker::ORI_SIGMA;
     const float DESC_SIGMA = cv::SURFInvoker::DESC_SIGMA;
     
     CvMat *sum = 0, *mask1 = 0, *mask_sum = 0;
